+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_data.py
@@ -12,13 +12,16 @@ from utils.column_values.categorical_column_values import (
     AscwdsFilteringRule,
     CareHome,
     CTFilteringRule,
+    CurrentCSSR,
     Dormancy,
     EstimateFilledPostsSource,
     JobGroupLabels,
     JobRoleFilteringRule,
     MainJobRoleLabels,
     PrimaryServiceType,
+    PrimaryServiceTypeSecondLevel,
     Region,
+    RUI,
     Sector,
 )
 
@@ -313,10 +316,10 @@ class ValidateModel01FeaturesData:
 @dataclass
 class MergeIndCQCData:
     cqc_location_data = [
-        ("1-001", date(2024, 1, 1), "Y", Sector.independent),
-        ("1-002", date(2024, 1, 1), "Y", Sector.local_authority),
-        ("1-003", date(2024, 1, 1), "N", Sector.independent),
-    ]
+        ("1-001", date(2024, 1, 1), "Y", Sector.independent, Dormancy.not_dormant, PrimaryServiceType.non_residential, PrimaryServiceTypeSecondLevel.non_residential, RUI.urban_major, Region.london, CurrentCSSR.barnet, "sub_icb_value", "icb_value", "icb_region_value", Region.eastern, CurrentCSSR.bedford, "icb_value"),
+        ("1-002", date(2024, 1, 1), "Y", Sector.local_authority, Dormancy.not_dormant, PrimaryServiceType.non_residential, PrimaryServiceTypeSecondLevel.non_residential, RUI.urban_major, Region.london, CurrentCSSR.barnet, "sub_icb_value", "icb_value", "icb_region_value", Region.eastern, CurrentCSSR.bedford, "icb_value"),
+        ("1-003", date(2024, 1, 1), "N", Sector.independent, Dormancy.not_dormant, PrimaryServiceType.non_residential, PrimaryServiceTypeSecondLevel.non_residential, RUI.urban_major, Region.london, CurrentCSSR.barnet, "sub_icb_value", "icb_value", "icb_region_value", Region.eastern, CurrentCSSR.bedford, "icb_value"),
+    ] # fmt: skip
     cqc_pir_data = [
         ("1-001", date(2024, 1, 1), "Y", "pir_value"),
         ("1-003", date(2024, 1, 1), "N", "pir_value"),
@@ -335,8 +338,8 @@ class MergeIndCQCData:
     ]
 
     expected_data = [
-        ("1-001", date(2024, 1, 1), "Y", Sector.independent, date(2024, 1, 1), "pir_value", date(2024, 1, 1), "ascwds_value", date(2024, 1, 1), "ct_non_res_value", date(2024, 1, 1), "ct_care_home_value"),
-        ("1-003", date(2024, 1, 1), "N", Sector.independent, date(2024, 1, 1), "pir_value", date(2024, 1, 1), "ascwds_value", date(2024, 1, 1), "ct_non_res_value", date(2024, 1, 1), "ct_care_home_value"),
+        ("1-001", date(2024, 1, 1), "Y", Sector.independent, Dormancy.not_dormant, PrimaryServiceType.non_residential, PrimaryServiceTypeSecondLevel.non_residential, RUI.urban_major, Region.london, CurrentCSSR.barnet, "sub_icb_value", "icb_value", "icb_region_value", Region.eastern, CurrentCSSR.bedford, "icb_value", date(2024, 1, 1), "pir_value", date(2024, 1, 1), "ascwds_value", date(2024, 1, 1), "ct_non_res_value", date(2024, 1, 1), "ct_care_home_value"),
+        ("1-003", date(2024, 1, 1), "N", Sector.independent, Dormancy.not_dormant, PrimaryServiceType.non_residential, PrimaryServiceTypeSecondLevel.non_residential, RUI.urban_major, Region.london, CurrentCSSR.barnet, "sub_icb_value", "icb_value", "icb_region_value", Region.eastern, CurrentCSSR.bedford, "icb_value", date(2024, 1, 1), "pir_value", date(2024, 1, 1), "ascwds_value", date(2024, 1, 1), "ct_non_res_value", date(2024, 1, 1), "ct_care_home_value"),
     ] # fmt: skip